feat: oficialmente testes de usuário terminados! Adicionados agora os testes de atualização de usuário e da senha, e exclusão

+++ src/main/java/com/project/ecopraia/service/UsuarioService.java
@@ -59,6 +59,11 @@ public class UsuarioService {
     }
 
     public void excluir(Long id) {
+
+        if (!usuarioRepository.existsById(id)) {
+            throw new RuntimeException("Usuário não encontrado para exclusão");
+        }
+
         usuarioRepository.deleteById(id);
     }
 }
